# openapi.yaml
openapi: 3.1.0
info:
  title: Pokemon Champions Helper API
  version: "1.0.10"
  description: Pokemon Champions OP.GG meta lookup, battle advice, team selection, and party building API.
servers:
  - url: https://pochamps-helper.onrender.com
paths:
  /health:
    get:
      operationId: checkHealth
      summary: Check API health
      responses:
        "200":
          description: OK
  /pokemon/{name}/advice:
    get:
      operationId: getPokemonAdvice
      summary: Get Pokemon Champions battle advice
      description: Returns representative estimated set, threat notes, counterplay, partners, and matchup notes based on OP.GG Pokemon Champions data.
      parameters:
        - name: name
          in: path
          required: true
          description: Pokemon Korean name, English name, or slug. Examples include 한카리아스, 메타그로스, garchomp, metagross.
          schema:
            type: string
        - name: refresh
          in: query
          required: false
          description: Whether to fetch live OP.GG data and update cache.
          schema:
            type: boolean
            default: false
        - name: battle_format
          in: query
          required: false
          description: Battle format. Use single for current MVP.
          schema:
            type: string
            default: single
      responses:
        "200":
          description: Battle advice response
          content:
            application/json:
              schema:
                type: object
  /pokemon/{name}/profile:
    get:
      operationId: getPokemonProfile
      summary: Get inferred role and type profile for one Pokemon
      description: Returns inferred type profile, role tags, weaknesses, resistances, immunities, and common attacking types.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: battle_format
          in: query
          required: false
          schema:
            type: string
            default: single
      responses:
        "200":
          description: Pokemon profile response
          content:
            application/json:
              schema:
                type: object
  /team/selection:
    post:
      operationId: recommendTeamSelection
      summary: Recommend 3 Pokemon selection from a 6 Pokemon party against opponent team
      description: Takes my_team and opponent_team, evaluates all 3-Pokemon combinations, and returns top recommended selections.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                my_team:
                  type: array
                  items:
                    type: string
                  description: My 3-6 Pokemon names. Usually 6 names.
                opponent_team:
                  type: array
                  items:
                    type: string
                  description: Opponent 1-6 Pokemon names. Usually 6 names.
                refresh:
                  type: boolean
                  default: false
                battle_format:
                  type: string
                  default: single
                top_n:
                  type: integer
                  default: 3
              required:
                - my_team
                - opponent_team
      responses:
        "200":
          description: Team selection recommendations
          content:
            application/json:
              schema:
                type: object
  /team/build:
    post:
      operationId: buildRecommendedParty
      summary: Build a recommended 6 Pokemon party around selected core Pokemon
      description: Takes 1-3 core Pokemon and suggests a party using OP.GG partners, weakness coverage, and role balance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                core:
                  type: array
                  items:
                    type: string
                  description: Core Pokemon names. Usually 1-3 names.
                candidate_pool:
                  type: array
                  items:
                    type: string
                  description: Optional Pokemon candidates to prefer.
                party_size:
                  type: integer
                  default: 6
                refresh:
                  type: boolean
                  default: false
                battle_format:
                  type: string
                  default: single
              required:
                - core
      responses:
        "200":
          description: Recommended party response
          content:
            application/json:
              schema:
                type: object
  /team/analyze:
    post:
      operationId: analyzeTeam
      summary: Analyze a Pokemon team
      description: Takes a team list and returns inferred role balance, weakness stacking, strengths, and warnings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                team:
                  type: array
                  items:
                    type: string
                refresh:
                  type: boolean
                  default: false
                battle_format:
                  type: string
                  default: single
              required:
                - team
      responses:
        "200":
          description: Team analysis response
          content:
            application/json:
              schema:
                type: object